Reorganize misc.py
