+++ tubesync/sync/models/__init__.py
@@ -1,16 +1,19 @@
 # These are referenced from the migration files
 
-from .misc import (
+from ._migrations import (
     get_media_file_path,
     get_media_thumb_path,
     media_file_storage,
 )
 
 # The actual model classes
+# The order starts with independent classes
+# then the classes that depend on them follow.
 
-from .media import Media
+from .media_server import MediaServer
+
 from .source import Source
+from .media import Media
 from .metadata import Metadata
 from .metadata_format import MetadataFormat
-from .media_server import MediaServer
